requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.5
idna==3.4
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
python-dateutil==2.8.2
razorpay==1.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
